moves page transition switcher to its own widget
